push for adaptor

+++ solvers/sat_rs/src/sat_tree.rs
@@ -1,15 +1,14 @@
-use rustsat::types::Lit;  
 use rustsat::instances::SatInstance;
+use rustsat::types::Lit;
 
-pub fn conv_to_clause(to_convert: &Vec<i16>, instance_in_use: &mut SatInstance) -> () {
+pub fn conv_to_clause(to_convert: &Vec<i32>, instance_in_use: &mut SatInstance) -> () {
     let l1: Lit = mk_lit(to_convert[0], instance_in_use);
     let l2: Lit = mk_lit(to_convert[1], instance_in_use);
 
     instance_in_use.add_binary(l1, l2); // (!x or y) and (x or y)
 }
 
-pub fn mk_lit(num: i16, instance_in_use: &mut SatInstance) -> Lit {
-
+pub fn mk_lit(num: i32, instance_in_use: &mut SatInstance) -> Lit {
     let var = instance_in_use.new_var();
     let lit;
 
@@ -23,8 +22,7 @@ pub fn mk_lit(num: i16, instance_in_use: &mut SatInstance) -> Lit {
     lit
 }
 
-pub fn conv_to_formula(vec_cnf: &Vec<Vec<i16>>, instance_in_use: &mut SatInstance) {
-    
+pub fn conv_to_formula(vec_cnf: &Vec<Vec<i32>>, instance_in_use: &mut SatInstance) {
     for value in vec_cnf {
         (conv_to_clause(value, instance_in_use));
     }
